Added Text, Cleaned up data.
Added text and described graphs and method. Also had to do another initial split for a subgroup so our earlier code was missing a step. The only thing I'm unsure about is the results visualization at the end. I think it is okay but I need to get it verified with a friend by tomorrow (Dec 5 as of now). We are just missing the discussion and conclusion section. I want to further elaborate on the confusion matrix and results visualization. Please feel free to make edits and extra explanations. Let me know in Instagram if you make any changes please and thank you!